feat: implement full authentication and role-based access control
- Update User entity and CustomUserDetailsService to map database roles to security authorities
- Configure SecurityConfig to protect book API endpoints with ADMIN authority
- Fix registration role binding and authentication data persistence
- Add Admin role validation to book management controllers

+++ backend/src/main/java/arl/repository/BookRepository.java
@@ -7,11 +7,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepository extends JpaRepository<Book, Long> {
-    
+public interface BookRepository extends JpaRepository<Book, Integer> { // FIXED: Changed Long to Integer
+
     /**
      * Student B Task: Search functionality.
-     * This method searches for books where the title contains the keyword 'q', 
+     * This method searches for books where the title contains the keyword 'title',
      * ignoring case, and returns a paginated result.
      */
     Page<Book> findByTitleContainingIgnoreCase(String title, Pageable pageable);
